moved the code to non-autogen file
